requirements.txt
# UI Framework
customtkinter

# AI Service Libraries
google-generativeai
requests

# Utilities
markdown-it-py
tkhtmlview
structlog
keyring
keyrings.alt
pydantic

# Testing Framework
pytest
pytest-mock